Removed create menu button and added additional checks to make sure menu is created for a restaurant

+++ RestaurantItem/models.py
@@ -4,7 +4,7 @@ from RestaurantLogin.models import Restaurant, RestaurantManager
 class Item(models.Model):
     item_title = models.CharField(max_length=255)
     item_description = models.TextField()
-    item_price = models.IntegerField()
+    item_price = models.DecimalField(max_digits = 10, decimal_places=2)
     # menus
     restaurant = models.ForeignKey(Restaurant,related_name="items", on_delete=models.CASCADE)
